Split mapping extensions into separate files
