+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/ChannelMapping.kt
@@ -1,0 +1,13 @@
+package io.getstream.chat.android.client.api2.mapping
+
+import io.getstream.chat.android.client.api2.model.dto.ChannelInfoDto
+import io.getstream.chat.android.client.models.ChannelInfo
+
+internal fun ChannelInfoDto.toDomain(): ChannelInfo =
+    ChannelInfo(
+        cid = cid,
+        id = id,
+        type = type,
+        memberCount = member_count,
+        name = name,
+    )
